[move] entity, repository 패키지 추가 및 기존 클래스 이동
related to #6
