+++ src/main/java/com/jo0oy/cafeorder/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.jo0oy.cafeorder.user;
+package com.jo0oy.cafeorder.user.repository;
 
+import com.jo0oy.cafeorder.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
